migrated to TAP testing
You will probably wonder why.

The answer is because the unittest module pulls in multiple other
dependencies and does various bits of syntax twiddling magic. And it
outputs a format that is slightly prettier but non-standard. TAP is
only pretty with a formatter, but there are numerous formatters to
choose from. There are also multiple test harnesses to choose from.
Anything can trivially output TAP and large projects that cross
multiple languages can run and report TAP tests as a single unified
whole.